refactor(CommentNode): move to separate file
